# SourceCode/front-end/logsmart/openapi.json
{
	"openapi": "3.1.0",
	"info": { "title": "back-end", "description": "", "license": { "name": "" }, "version": "1.1.0" },
	"paths": {
...
		"/logs/templates": {
			"get": {
				"tags": ["Templates"],
				"summary": "Retrieves a specific log template by name.",
				"operationId": "get_template",
				"parameters": [
					{
						"name": "template_name",
						"in": "query",
						"description": "Name of the template to retrieve",
						"required": true,
						"schema": { "type": "string" },
						"example": "ErrorLog"
					}
				],
				"responses": {
					"200": {
						"description": "Template retrieved successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/GetTemplateResponse" }
							}
						}
					},
					"401": {
						"description": "Invalid or expired token",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"404": {
						"description": "Template not found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			},
			"post": {
				"tags": ["Templates"],
				"summary": "Adds a new log template for the current company.",
				"operationId": "add_template",
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/AddTemplateRequest" } }
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Template added successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/AddTemplateResponse" }
							}
						}
					},
					"401": {
						"description": "Invalid or expired token",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			},
			"delete": {
				"tags": ["Templates"],
				"summary": "Deletes a specific log template.",
				"operationId": "delete_template",
				"parameters": [
					{
						"name": "template_name",
						"in": "query",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "Template deleted successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/DeleteTemplateResponse" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
		"/logs/templates/all": {
			"get": {
				"tags": ["Templates"],
				"summary": "Retrieves all log templates for the current company/branch.",
				"operationId": "get_all_templates",
				"responses": {
					"200": {
						"description": "All templates retrieved successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/GetAllTemplatesResponse" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
		"/logs/templates/rename": {
			"put": {
				"tags": ["Templates"],
				"summary": "Renames an existing log template.",
				"operationId": "rename_template",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/RenameTemplateRequest" }
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Template renamed successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/RenameTemplateResponse" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
		"/logs/templates/update": {
			"put": {
				"tags": ["Templates"],
				"summary": "Updates an existing log template.",
				"operationId": "update_template",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/UpdateTemplateRequest" }
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Template updated successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/UpdateTemplateResponse" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
		"/logs/templates/versions": {
			"get": {
				"tags": ["Templates"],
				"summary": "Retrieves the version history of a log template.",
				"operationId": "get_template_versions",
				"parameters": [
					{
						"name": "template_name",
						"in": "query",
						"description": "Name of the template to retrieve versions for",
						"required": true,
						"schema": { "type": "string" },
						"example": "ErrorLog"
					}
				],
				"responses": {
					"200": {
						"description": "Versions retrieved successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/GetTemplateVersionsResponse" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
		"/logs/templates/versions/restore": {
			"post": {
				"tags": ["Templates"],
				"summary": "Restores a specific version of a log template.",
				"operationId": "restore_template_version",
				"parameters": [
					{
						"name": "template_name",
						"in": "query",
						"description": "Name of the template to restore",
						"required": true,
						"schema": { "type": "string" },
						"example": "ErrorLog"
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/RestoreTemplateVersionRequest" }
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Template restored successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/UpdateTemplateResponse" }
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"AddTemplateRequest": {
				"type": "object",
				"required": ["template_name", "template_layout", "schedule"],
				"properties": {
					"branch_id": { "type": ["string", "null"], "example": "branch-uuid-here" },
					"schedule": { "$ref": "#/components/schemas/Schedule" },
					"template_layout": { "$ref": "#/components/schemas/Vec" },
					"template_name": { "type": "string", "example": "Kitchen Daily Log" }
				}
			},
			"AddTemplateResponse": {
				"type": "object",
				"required": ["message"],
				"properties": { "message": { "type": "string" } }
			},
...
			"DeleteTemplateResponse": {
				"type": "object",
				"required": ["message"],
				"properties": { "message": { "type": "string" } }
			},
...
			"ErrorResponse": {
				"type": "object",
				"required": ["error"],
				"properties": {
					"error": { "type": "string" },
					"fields": {
						"type": ["object", "null"],
						"additionalProperties": { "type": "array", "items": { "type": "string" } },
						"propertyNames": { "type": "string" }
					}
				}
			},
			"Frequency": {
				"type": "string",
				"enum": ["Daily", "Weekly", "Monthly", "Quarterly", "Yearly"]
			},
			"GetAllTemplatesResponse": {
				"type": "object",
				"required": ["templates"],
				"properties": {
					"templates": { "type": "array", "items": { "$ref": "#/components/schemas/TemplateInfo" } }
				}
			},
...
			"GetTemplateResponse": {
				"type": "object",
				"required": ["template_name", "template_layout", "version"],
				"properties": {
					"branch_id": { "type": ["string", "null"] },
					"template_layout": { "$ref": "#/components/schemas/Vec" },
					"template_name": { "type": "string" },
					"version": { "type": "integer", "format": "int32", "minimum": 0 },
					"version_name": { "type": ["string", "null"] }
				}
			},
			"GetTemplateVersionsResponse": {
				"type": "object",
				"required": ["versions"],
				"properties": {
					"versions": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/TemplateVersionInfo" }
					}
				}
			},
...
			"Position": {
				"type": "object",
				"required": ["x", "y"],
				"properties": {
					"x": { "type": "number", "format": "double" },
					"y": { "type": "number", "format": "double" }
				}
			},
...
			"RenameTemplateRequest": {
				"type": "object",
				"required": ["old_template_name", "new_template_name"],
				"properties": {
					"new_template_name": { "type": "string" },
					"old_template_name": { "type": "string" }
				}
			},
			"RenameTemplateResponse": {
				"type": "object",
				"required": ["message"],
				"properties": { "message": { "type": "string" } }
			},
...
			"RestoreTemplateVersionRequest": {
				"type": "object",
				"required": ["version"],
				"properties": { "version": { "type": "integer", "format": "int32", "minimum": 0 } }
			},
			"Schedule": {
				"type": "object",
				"required": ["frequency"],
				"properties": {
					"available_from_time": { "type": ["string", "null"] },
					"day_of_month": { "type": ["integer", "null"], "format": "int32", "minimum": 0 },
					"day_of_week": { "type": ["integer", "null"], "format": "int32", "minimum": 0 },
					"days_of_week": {
						"type": ["array", "null"],
						"items": { "type": "integer", "format": "int32", "minimum": 0 }
					},
					"due_at_time": { "type": ["string", "null"] },
					"frequency": { "$ref": "#/components/schemas/Frequency" },
					"month_of_year": { "type": ["integer", "null"], "format": "int32", "minimum": 0 }
				}
			},
...
			"TemplateFieldProps": {
				"type": "object",
				"properties": {
					"color": { "type": ["string", "null"] },
					"editable": { "type": ["boolean", "null"] },
					"font_family": { "type": ["string", "null"] },
					"input_type": { "type": ["string", "null"] },
					"max": { "type": ["number", "null"], "format": "float" },
					"max_length": { "type": ["integer", "null"], "format": "int32" },
					"min": { "type": ["number", "null"], "format": "float" },
					"min_length": { "type": ["integer", "null"], "format": "int32" },
					"options": { "type": ["array", "null"], "items": { "type": "string" } },
					"placeholder": { "type": ["string", "null"] },
					"required": { "type": ["boolean", "null"] },
					"selected": { "type": ["string", "null"] },
					"size": { "type": ["string", "null"] },
					"text": { "type": ["string", "null"] },
					"text_decoration": { "type": ["string", "null"] },
					"unit": { "type": ["string", "null"] },
					"value": { "type": ["string", "null"] },
					"weight": { "type": ["string", "null"] }
				}
			},
			"TemplateInfo": {
				"type": "object",
				"required": ["template_name", "schedule", "created_by", "created_at", "updated_at"],
				"properties": {
					"created_at": { "type": "string" },
					"created_by": { "type": "string" },
					"schedule": { "$ref": "#/components/schemas/Schedule" },
					"template_name": { "type": "string" },
					"updated_at": { "type": "string" }
				}
			},
			"TemplateVersionInfo": {
				"type": "object",
				"required": ["version", "created_at", "created_by"],
				"properties": {
					"created_at": { "type": "string" },
					"created_by": { "type": "string" },
					"version": { "type": "integer", "format": "int32", "minimum": 0 },
					"version_name": { "type": ["string", "null"] }
				}
			},
...
			"UpdateTemplateRequest": {
				"type": "object",
				"required": ["template_name"],
				"properties": {
					"branch_id": { "type": ["string", "null"], "example": "branch-uuid-here" },
					"schedule": {
						"oneOf": [{ "type": "null" }, { "$ref": "#/components/schemas/Schedule" }]
					},
					"template_layout": {
						"oneOf": [{ "type": "null" }, { "$ref": "#/components/schemas/Vec" }]
					},
					"template_name": { "type": "string", "example": "Kitchen Daily Log" },
					"version_name": { "type": ["string", "null"], "example": "Major Update" }
				}
			},
			"UpdateTemplateResponse": {
				"type": "object",
				"required": ["message"],
				"properties": { "message": { "type": "string" } }
			},
...
			"Vec": {
				"type": "array",
				"items": {
					"type": "object",
					"required": ["field_type", "position", "props"],
					"properties": {
						"field_type": { "type": "string" },
						"position": { "$ref": "#/components/schemas/Position" },
						"props": { "$ref": "#/components/schemas/TemplateFieldProps" }
					}
				}
			},
...
		},
		"securitySchemes": { "bearer_auth": { "type": "http", "scheme": "bearer" } }
	},
...
}